Module Problems__First_ten

Functions for the first ten problems

val last : 'a list -> 'a option

last returns the last element of a list

val last_two : 'a list -> ('a * 'a) option

last_two returns the last two elemetns of a list as a tuple

exception Invalid_k of string

val at : int -> 'a list -> 'a

at returns kth element of a list

raises Invalid_k

if k is less than

raises Failure

if list is out of bounds for k

val length : 'a list -> int

length returns length of a list. naive implementation

val length' : 'a list -> int

length' returns length of a list. tail recursive

val rev : 'a list -> 'a list

rev returns a list reversed. naive implementaion

val rev' : 'a list -> 'a list

rev' returns a list reversed. tail recursive

val is_palindrome : 'a list -> bool

is_palindrome returns if list is true or false

type 'a node =

| One of 'a

| Many of 'a node list

val flatten : 'a node list -> 'a list

flatten returns flattened nested list

val compress : 'a list -> 'a list

compress returns a the list passed with consecutives duplicates removed

val pack : 'a list -> 'a list list

pack returns a list with sublists packed from the passed list of consecutive duplicates

val encode : 'a list -> (int * 'a) list

encode returns the run-length encoding of the passed list