Module Problems__Second_ten

Functions for the second set of ten problems

exception Invalid_argument of string

type 'a rle =

| One of 'a

| Many of int * 'a

val encode' : 'a list -> 'a rle list

encode' returns the run-length encoding of the passed list using the type 'a rle

val decode : 'a rle list -> 'a list

decode returns an uncompressed version rle

val encode'' : 'a list -> 'a rle list

encode'' returns the run-length encoding of the passed list using the type 'a rle Same as encode'.

val duplicate : 'a list -> 'a list

duplicate takes a list and returns with each element duplicated

val replicate : 'a list -> int -> 'a list

replicate takes a list and times and returns a new list with each element in list replicated times

val drop : 'a list -> int -> 'a list

drop returns a new list with every nth element dropped from the original passed list

val split : 'a list -> int -> 'a list * 'a list

split returns a split pair from list lst passed with split at n

val slice : 'a list -> int -> int -> 'a list

slice takes two indices i k and returns a slice from list lst. List is zero indexed. i and k should be positive and i <= k

val rotate : 'a list -> int -> 'a list

rotate takes a list lst and rotates it to left by n places

To rotate by right use negative n

Raises: Invalid_argument if abs of n is greater than length of lst

val remove_at : int -> 'a list -> 'a list

remove_at takes a list lst and removes the element at index from the list